Rename `query_impl.rs` as `query_vtables.rs`
The current name is very generic and the file is mostly about query
vtables.
