+++ compiler/rustc_query_impl/src/execution.rs
@@ -21,7 +21,7 @@ use tracing::debug;
 use crate::handle_cycle_error;
 use crate::job::{QueryJobInfo, QueryJobMap, find_cycle_in_stack};
 use crate::plumbing::{current_query_job, next_job_id, start_query};
-use crate::query_impl::for_each_query_vtable;
+use crate::query_vtables::for_each_query_vtable;
 
 #[inline]
 fn equivalent_key<K: Eq, V>(k: K) -> impl Fn(&(K, V)) -> bool {
